+++ packages/swc-plugin/transform/tests/fixture.rs
@@ -4,13 +4,13 @@ use swc_ecma_parser::{Syntax, TsSyntax};
 use swc_ecma_transforms_testing::test_fixture;
 use swc_global_module::global_module;
 
-const MODULE_ID: u32 = 1000;
+const MODULE_ID: f64 = 1000.0;
 
 #[testing::fixture("tests/fixture/register/**/input.js")]
 fn register_fixture(input: PathBuf) {
     let filename = input.to_string_lossy();
     let output = input.with_file_name("output.js");
-    let phase = 0; // ModulePhase::Register
+    let phase = 0.0; // ModulePhase::Register
 
     test_fixture(
         Syntax::Typescript(TsSyntax {
@@ -28,7 +28,7 @@ fn register_fixture(input: PathBuf) {
 fn runtime_fixture(input: PathBuf) {
     let filename = input.to_string_lossy();
     let output = input.with_file_name("output.js");
-    let phase = 1; // ModulePhase::Runtime
+    let phase = 1.0; // ModulePhase::Runtime
 
     test_fixture(
         Syntax::Typescript(TsSyntax {
